+++ samples/SampleState/Program.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using SampleState.State.Reducers;
 using NetRx.Store;
-using System.Linq;
 using actions = SampleState.State.Actions;
+using System.Collections.Immutable;
 
 namespace SampleState
 {
@@ -15,7 +14,7 @@ namespace SampleState
             {
                 Email = string.Empty,
                 Name = string.Empty,
-                Contacts = new List<Contact>()
+                Contacts = ImmutableArray.Create<Contact>()
             };
 
             var store = Store.Create().WithState(initialState, new ProfileReducer());
@@ -28,13 +27,13 @@ namespace SampleState
             {
                 Console.WriteLine($"Email: {value}");
             });
-            store.Select<ProfileState, List<Contact>>(state => state.Contacts).Subscribe(value =>
+            store.Select<ProfileState, ImmutableArray<Contact>> (state => state.Contacts).Subscribe(value =>
             {
-                Console.WriteLine($"Contacts: {value?.Count()}");
+                Console.WriteLine($"Contacts: {value.Length}");
             });
             store.Select<ProfileState, ProfileState>(state => state).Subscribe(value =>
             {
-                Console.WriteLine($"State value: {value.Name} {value.Email} {value.Contacts?.Count()}");
+                Console.WriteLine($"State value: {value.Name} {value.Email} {value.Contacts.Length}");
             });
 
             store.Dispatch(new actions.SetEmail("test@mail.com"));