+++ samples/SampleState/State/Reducers/ProfileReducer.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Immutable;
 using System.Linq;
 using NetRx.Store;
 using actions = SampleState.State.Actions;
@@ -19,7 +18,7 @@ namespace SampleState.State.Reducers
 
         public string Email { get; set; }
 
-        public List<Contact> Contacts { get; set; }
+        public ImmutableArray<Contact> Contacts { get; set; }
     }
 
     public class ProfileReducer : Reducer<ProfileState>
@@ -38,13 +37,13 @@ namespace SampleState.State.Reducers
 
         public ProfileState Reduce(ProfileState state, actions.AddContact action)
         {
-            state.Contacts = state.Contacts.Concat(new [] { action.Payload }).ToList();
+            state.Contacts = state.Contacts.Concat(new [] { action.Payload }).ToImmutableArray();
             return state;
         }
 
         public ProfileState Reduce(ProfileState state, actions.ClearContacts action)
         {
-            state.Contacts = new List<Contact>();
+            state.Contacts = ImmutableArray.Create<Contact>();
             return state;
         }
     }